requirements.txt
Fabric3==1.13.1.post1

SQLAlchemy==1.2.7
#  implicit dependency
#  usage: DB/ORM library
#  repo: https://github.com/zzzeek/sqlalchemy
#  license: MIT - <repo>/blob/master/LICENSE

SQLAlchemy-Utils==0.32.19
#  usage: for nice custom fields (e.g., PasswordField) and database helpers
#  repo: https://github.com/kvesteri/sqlalchemy-utils
#  license: BSD (3-Clause) - <repo>/blob/master/LICENSE